pipeline/requirements.txt
openai==1.68.2
requests==2.32.3
feedparser==6.0.11
python-slugify==8.0.4
finnhub-python==2.4.21
websocket-client==1.8.0
